requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic-settings
python-dotenv
crawlee[playwright]
playwright
anthropic
openai
httpx
apscheduler
typer[all]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-multipart
